# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  name: string;
  avatarColor: string;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  avatarColor?: string;
}

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  avatarColor?: string;
}

export type ChoreFrequency = typeof ChoreFrequency[keyof typeof ChoreFrequency];


export const ChoreFrequency = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export interface Chore {
  id: number;
  name: string;
  icon: string;
  dollarValue: number;
  frequency: ChoreFrequency;
  assignedMemberIds?: number[];
  createdAt: string;
}

export type ChoreInputFrequency = typeof ChoreInputFrequency[keyof typeof ChoreInputFrequency];


export const ChoreInputFrequency = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export interface ChoreInput {
  /** @minLength 1 */
  name: string;
  icon: string;
  /** @minimum 0 */
  dollarValue: number;
  frequency: ChoreInputFrequency;
}

export type ChoreUpdateFrequency = typeof ChoreUpdateFrequency[keyof typeof ChoreUpdateFrequency];


export const ChoreUpdateFrequency = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export interface ChoreUpdate {
  /** @minLength 1 */
  name?: string;
  icon?: string;
  /** @minimum 0 */
  dollarValue?: number;
  frequency?: ChoreUpdateFrequency;
}

export interface Assignment {
  id: number;
  choreId: number;
  memberId: number;
}

export interface AssignmentInput {
  memberId: number;
}

export interface ChecklistItem {
  chore: Chore;
  /** @nullable */
  completionId: number | null;
  /** @nullable */
  completedAt: string | null;
}

export interface MemberChecklist {
  member: Member;
  chores: ChecklistItem[];
}

export interface Completion {
  id: number;
  choreId: number;
  memberId: number;
  amountEarned: number;
  completedAt: string;
}

export interface CompletionInput {
  choreId: number;
  memberId: number;
}

export interface HistoryEntry {
  id: number;
  choreId: number;
  choreName: string;
  choreIcon: string;
  memberId: number;
  memberName: string;
  memberAvatarColor: string;
  amountEarned: number;
  completedAt: string;
}

export interface HistoryResponse {
  entries: HistoryEntry[];
  total: number;
}

export interface MemberBalance {
  member: Member;
  balance: number;
  totalEarned: number;
  totalPaidOut: number;
}

export interface Payout {
  id: number;
  memberId: number;
  memberName: string;
  amount: number;
  paidAt: string;
}

export interface PayoutInput {
  memberId: number;
  /** @minimum 0 */
  amount: number;
}

export interface HouseholdSummary {
  totalEarnedAllTime: number;
  totalChoresCompletedToday: number;
  totalChoresCompletedThisWeek: number;
  /** @nullable */
  topEarnerName: string | null;
  /** @nullable */
  topEarnerBalance: number | null;
  totalOutstandingBalance: number;
}

export type GetHistoryParams = {
/**
 * @nullable
 */
memberId?: number | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};