# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */
import type { Chore } from './chore';

export interface ChecklistItem {
  chore: Chore;
  /** @nullable */
  completionId: number | null;
  /** @nullable */
  completedAt: string | null;
}