# lib/api-zod/src/generated/types/chore.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */
import type { ChoreFrequency } from './choreFrequency';

export interface Chore {
  id: number;
  name: string;
  icon: string;
  dollarValue: number;
  frequency: ChoreFrequency;
  assignedMemberIds?: number[];
  createdAt: string;
}