# lib/api-zod/src/generated/types/choreInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */
import type { ChoreInputFrequency } from './choreInputFrequency';

export interface ChoreInput {
  /** @minLength 1 */
  name: string;
  icon: string;
  /** @minimum 0 */
  dollarValue: number;
  frequency: ChoreInputFrequency;
}