# lib/api-zod/src/generated/types/completion.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */

export interface Completion {
  id: number;
  choreId: number;
  memberId: number;
  amountEarned: number;
  completedAt: string;
}