# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  choreId: number;
  choreName: string;
  choreIcon: string;
  memberId: number;
  memberName: string;
  memberAvatarColor: string;
  amountEarned: number;
  completedAt: string;
}