# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntry } from './historyEntry';

export interface HistoryResponse {
  entries: HistoryEntry[];
  total: number;
}