# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Chorific API - Household chore chart and allowance tracker
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  avatarColor?: string;
}